backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# CORS
python-multipart==0.0.6

# MCP SDK
mcp>=0.9.0

# Google Gemini SDK (replacing OpenAI)
google-generativeai>=0.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
